addressed PR feedback
