+++ pkg/ruler/querier_test.go
@@ -1,0 +1,74 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package ruler
+
+import (
+	"context"
+	"net/http"
+	"testing"
+	"time"
+
+	"github.com/go-kit/log"
+	"github.com/gogo/protobuf/proto"
+	"github.com/golang/snappy"
+	"github.com/prometheus/prometheus/prompb"
+	"github.com/stretchr/testify/require"
+	"github.com/weaveworks/common/httpgrpc"
+	"google.golang.org/grpc"
+)
+
+type mockHTTPGrpcClient func(ctx context.Context, req *httpgrpc.HTTPRequest, _ ...grpc.CallOption) (*httpgrpc.HTTPResponse, error)
+
+func (c mockHTTPGrpcClient) Handle(ctx context.Context, req *httpgrpc.HTTPRequest, opts ...grpc.CallOption) (*httpgrpc.HTTPResponse, error) {
+	return c(ctx, req, opts...)
+}
+
+func TestQuerier_ReadReq(t *testing.T) {
+	var inReq *httpgrpc.HTTPRequest
+	var body []byte
+
+	mockClientFn := func(ctx context.Context, req *httpgrpc.HTTPRequest, _ ...grpc.CallOption) (*httpgrpc.HTTPResponse, error) {
+		inReq = req
+
+		b, err := proto.Marshal(&prompb.ReadResponse{
+			Results: []*prompb.QueryResult{
+				{},
+			},
+		})
+		require.NoError(t, err)
+
+		body = snappy.Encode(nil, b)
+		return &httpgrpc.HTTPResponse{
+			Code: http.StatusOK,
+			Body: snappy.Encode(nil, b),
+		}, nil
+	}
+	q := New(mockHTTPGrpcClient(mockClientFn), "/prometheus", log.NewNopLogger())
+
+	_, err := q.Read(context.Background(), &prompb.Query{})
+	require.NoError(t, err)
+
+	require.NotNil(t, inReq)
+	require.Equal(t, http.MethodPost, inReq.Method)
+	require.Equal(t, body, inReq.Body)
+	require.Equal(t, "/prometheus/api/v1/read", inReq.Url)
+}
+
+func TestQuerier_QueryReq(t *testing.T) {
+	var inReq *httpgrpc.HTTPRequest
+	mockClientFn := func(ctx context.Context, req *httpgrpc.HTTPRequest, _ ...grpc.CallOption) (*httpgrpc.HTTPResponse, error) {
+		inReq = req
+		return &httpgrpc.HTTPResponse{Code: http.StatusOK, Body: []byte(`{
+							"status": "success","data": {"resultType":"vector","result":[]}
+						}`)}, nil
+	}
+	q := New(mockHTTPGrpcClient(mockClientFn), "/prometheus", log.NewNopLogger())
+
+	_, _, err := q.Query(context.Background(), "qs", time.Unix(1649092025, 515834))
+	require.NoError(t, err)
+
+	require.NotNil(t, inReq)
+	require.Equal(t, http.MethodPost, inReq.Method)
+	require.Equal(t, "query=qs&time=1649092025.000516", string(inReq.Body))
+	require.Equal(t, "/prometheus/api/v1/query", inReq.Url)
+}
